requirements.txt
rapidfuzz
fastapi
uvicorn
google-generativeai
python-dotenv
supabase
numpy
setuptools
PyJWT[crypto]
requests
spacy